requirements.txt
alabaster==0.7.16
alembic==1.17.0
altair==5.5.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
atpublic==6.0.2
attrs==25.4.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.14.2
black==20.8b1
bleach==6.2.0
blinker==1.9.0
blosc2==3.10.2
boto3==1.40.55
botocore==1.40.55
bump2version==1.0.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
choreographer==1.2.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.2
coverage==7.11.0
cycler==0.12.1
dash==3.2.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
distlib==0.4.0
distro==1.9.0
docutils==0.16
docx==0.2.4
dpath==2.2.0
dulwich==0.24.6
dvc==1.10.1
dynaconf==2.2.3
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastcore==1.8.13
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
flatten-dict==0.4.2
flufl.lock==3.2
fonttools==4.60.1
fqdn==1.5.1
funcy==2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.6
greenlet==3.2.4
h11==0.16.0
html2image==2.0.7
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.0.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==4.3.21
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==1.1.0
kiwisolver==1.4.9
lark==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
logistro==2.0.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
msgpack==1.1.2
mypy_extensions==1.1.0
nanotime==0.5.2
narwhals==2.8.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==2.4
nodeenv==1.9.1
notebook==7.4.7
notebook_shim==0.2.4
numba==0.62.1
numexpr==2.14.1
numpy==1.26.4
onnx==1.19.1
openpyxl==3.1.5
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
plotly-resampler==0.11.0
pluggy==1.6.0
ply==3.11
pooch==1.8.2
pre-commit==2.21.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==1.10.24
pydeck==0.9.1
pydocstyle==6.3.0
pydot==4.0.1
pydub==0.25.1
pyflakes==2.5.0
Pygments==2.19.2
pygtrie==2.3.2
pylama==7.7.1
pyparsing==3.2.5
pytest==8.4.2
pytest-timeout==2.4.0
python-benedict==0.35.0
python-box==3.4.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-fsutil==0.15.0
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.3
pyzmq==27.1.0
rarfile==4.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.3
shapely==2.1.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
Sphinx==1.8.6
sphinx-autodoc-typehints==1.6.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-websupport==1.2.4
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.42.0
sqlalchemy_schemadisplay==1.4
stack-data==0.6.3
stqdm==0.0.5
streamlit==1.50.0
streamlit-jupyter==0.3.1
tables==3.10.1
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
tox==3.28.0
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typed-ast==1.5.5
typer==0.19.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
useful_types==0.2.1
virtualenv==20.35.3
voluptuous==0.15.2
watchdog==6.0.0
wavely.signal @ git+ssh://git@github.com/Wavely/signal.git@master#egg=wavely.signal
wavely.edge-metadata @ git+ssh://git@github.com/Wavely/edge-metadata.git@master#egg=wavely.edge-metadata
wavely.metadata @ git+ssh://git@github.com/Wavely/metadata.git@master#egg=wavely.metadata
wavely.datasets-analysis @ git+ssh://git@github.com/Wavely/datasets-analysis.git@master#egg=wavely.datasets-analysis
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xlrd==1.2.0
zc.lockfile==4.0
zipp==3.23.0
